Fix activity migration

+++ src/main/resources/db/migration/V6__Add_activity.sql
@@ -3,10 +3,10 @@ CREATE TABLE IF NOT EXISTS activity
     id          BIGINT AUTO_INCREMENT PRIMARY KEY,
     name        VARCHAR(255) NOT NULL,
     price       DOUBLE       NOT NULL,
-    duration    DECIMAL      NOT NULL,
+    duration    BIGINT       NOT NULL,
     provider_id BIGINT       NOT NULL,
     constraint fk_activity_provider
-        FOREIGN KEY (provider_id) references provider (id) ON DELETE CASCADE
+        FOREIGN KEY (provider_id) references provider(id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS activity_employees
@@ -16,5 +16,5 @@ CREATE TABLE IF NOT EXISTS activity_employees
     constraint fk_activity_id
         FOREIGN KEY (activity_id) references activity (id) ON DELETE CASCADE,
     constraint fk_employee_id
-        FOREIGN KEY (employee_id) references employee (id)
+        FOREIGN KEY (employee_id) references employee(id)
 );
